stack: Move stack command and subcommands up one level
