+++ cmd/stack/delete.go
@@ -15,22 +15,32 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package cmd
+package cmdstack
 
 import (
-	cmdauth "github.com/elastic/ecctl/cmd/auth"
-	cmddeployment "github.com/elastic/ecctl/cmd/deployment"
-	cmdplatform "github.com/elastic/ecctl/cmd/platform"
-	cmdstack "github.com/elastic/ecctl/cmd/stack"
-	cmduser "github.com/elastic/ecctl/cmd/user"
+	"github.com/elastic/cloud-sdk-go/pkg/api/stackapi"
+	"github.com/spf13/cobra"
+
+	cmdutil "github.com/elastic/ecctl/cmd/util"
+	"github.com/elastic/ecctl/pkg/ecctl"
 )
 
+var stackDeleteCmd = &cobra.Command{
+	Use:     "delete",
+	Short:   cmdutil.AdminReqDescription("Deletes an Elastic StackPack"),
+	PreRunE: cobra.MinimumNArgs(1),
+
+	RunE: func(cmd *cobra.Command, args []string) error {
+		return stackapi.Delete(stackapi.DeleteParams{
+			API:     ecctl.Get().API,
+			Region:  ecctl.Get().Config.Region,
+			Version: args[0],
+		})
+	},
+}
+
 func init() {
-	RootCmd.AddCommand(
-		cmdauth.Command,
-		cmddeployment.Command,
-		cmdplatform.Command,
-		cmduser.Command,
-		cmdstack.Command,
+	Command.AddCommand(
+		stackDeleteCmd,
 	)
 }
