+++ cmd/stack/upload.go
@@ -15,22 +15,40 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package cmd
+package cmdstack
 
 import (
-	cmdauth "github.com/elastic/ecctl/cmd/auth"
-	cmddeployment "github.com/elastic/ecctl/cmd/deployment"
-	cmdplatform "github.com/elastic/ecctl/cmd/platform"
-	cmdstack "github.com/elastic/ecctl/cmd/stack"
-	cmduser "github.com/elastic/ecctl/cmd/user"
+	"os"
+
+	"github.com/elastic/cloud-sdk-go/pkg/api/stackapi"
+	"github.com/spf13/cobra"
+
+	cmdutil "github.com/elastic/ecctl/cmd/util"
+	"github.com/elastic/ecctl/pkg/ecctl"
 )
 
+var stackUploadCmd = &cobra.Command{
+	Use:     "upload",
+	Short:   cmdutil.AdminReqDescription("Uploads an Elastic StackPack"),
+	PreRunE: cobra.MinimumNArgs(1),
+
+	RunE: func(cmd *cobra.Command, args []string) error {
+		f, err := os.Open(args[0])
+		if err != nil {
+			return err
+		}
+		defer f.Close()
+
+		return stackapi.Upload(stackapi.UploadParams{
+			API:       ecctl.Get().API,
+			Region:    ecctl.Get().Config.Region,
+			StackPack: f,
+		})
+	},
+}
+
 func init() {
-	RootCmd.AddCommand(
-		cmdauth.Command,
-		cmddeployment.Command,
-		cmdplatform.Command,
-		cmduser.Command,
-		cmdstack.Command,
+	Command.AddCommand(
+		stackUploadCmd,
 	)
 }
