+++ docs/ecctl-command-reference-index.adoc
@@ -94,11 +94,11 @@ include::ecctl_platform_runner_list.adoc[]
 include::ecctl_platform_runner_resync.adoc[]
 include::ecctl_platform_runner_search.adoc[]
 include::ecctl_platform_runner_show.adoc[]
-include::ecctl_platform_stack.adoc[]
-include::ecctl_platform_stack_delete.adoc[]
-include::ecctl_platform_stack_list.adoc[]
-include::ecctl_platform_stack_show.adoc[]
-include::ecctl_platform_stack_upload.adoc[]
+include::ecctl_stack.adoc[]
+include::ecctl_stack_delete.adoc[]
+include::ecctl_stack_list.adoc[]
+include::ecctl_stack_show.adoc[]
+include::ecctl_stack_upload.adoc[]
 include::ecctl_user.adoc[]
 include::ecctl_user_create.adoc[]
 include::ecctl_user_delete.adoc[]
