+++ docs/ecctl.adoc
@@ -39,5 +39,6 @@ Elastic Cloud Control
 * xref:ecctl_generate[ecctl generate]	 - Generates completions and docs
 * xref:ecctl_init[ecctl init]	 - Creates an initial configuration file.
 * xref:ecctl_platform[ecctl platform]	 - Manages the platform
+* xref:ecctl_stack[ecctl stack]	 - Manages Elastic StackPacks
 * xref:ecctl_user[ecctl user]	 - Manages the platform users (Available for ECE only)
 * xref:ecctl_version[ecctl version]	 - Shows ecctl version
