+++ docs/ecctl.md
@@ -35,6 +35,7 @@ Elastic Cloud Control
 * [ecctl generate](ecctl_generate.md)	 - Generates completions and docs
 * [ecctl init](ecctl_init.md)	 - Creates an initial configuration file.
 * [ecctl platform](ecctl_platform.md)	 - Manages the platform
+* [ecctl stack](ecctl_stack.md)	 - Manages Elastic StackPacks
 * [ecctl user](ecctl_user.md)	 - Manages the platform users (Available for ECE only)
 * [ecctl version](ecctl_version.md)	 - Shows ecctl version
