+++ docs/ecctl_stack_list.adoc
@@ -1,24 +1,29 @@
-## ecctl platform stack
+[#ecctl_stack_list]
+== ecctl stack list
 
-Manages Elastic StackPacks
+Lists Elastic StackPacks
 
-### Synopsis
+[float]
+=== Synopsis
 
-Manages Elastic StackPacks
+Lists Elastic StackPacks
 
-```
-ecctl platform stack [flags]
-```
+----
+ecctl stack list [flags]
+----
 
-### Options
+[float]
+=== Options
 
-```
-  -h, --help   help for stack
-```
+----
+  -d, --deleted   Shows deleted stackpacks
+  -h, --help      help for list
+----
 
-### Options inherited from parent commands
+[float]
+=== Options inherited from parent commands
 
-```
+----
       --apikey string      API key to use to authenticate (If empty will look for EC_APIKEY environment variable)
       --config string      Config name, used to have multiple configs in $HOME/.ecctl/<env> (default "config")
       --force              Do not ask for confirmation
@@ -35,13 +40,9 @@ ecctl platform stack [flags]
       --trace              Enables tracing saves the trace to trace-20060102150405
       --user string        Username to use to authenticate (If empty will look for EC_USER environment variable)
       --verbose            Enable verbose mode
-```
+----
 
-### SEE ALSO
+[float]
+=== SEE ALSO
 
-* [ecctl platform](ecctl_platform.md)	 - Manages the platform
-* [ecctl platform stack delete](ecctl_platform_stack_delete.md)	 - Deletes an Elastic StackPack (Available for ECE only)
-* [ecctl platform stack list](ecctl_platform_stack_list.md)	 - Lists Elastic StackPacks
-* [ecctl platform stack show](ecctl_platform_stack_show.md)	 - Shows information about an Elastic StackPack
-* [ecctl platform stack upload](ecctl_platform_stack_upload.md)	 - Uploads an Elastic StackPack (Available for ECE only)
-
+* xref:ecctl_stack[ecctl stack]	 - Manages Elastic StackPacks
